weather_client.py
import asyncio
from mcp.client.session import ClientSession
from mcp.client.stdio import stdio_client
from mcp import StdioServerParameters


async def get_weather_from_mcp(city: str):
    """
    Connects to the weather_mcp server and fetches weather for the given city.
    """
    server_params = StdioServerParameters(command="python", args=["weather_server.py"])
    async with stdio_client(server_params) as (reader, writer):
        async with ClientSession(reader, writer) as session:
            await session.initialize()
            result = await session.call_tool("get_weather", {"location": city})
            return result
