requirements.txt
fastmcp
requests
python-dotenv
transformers
torch